feat: Export an order in Excel file from a plain file
This is the core feature of the script. It takes a plain file with the
name of the buyer on the first row and the list of the items to order with
the quantity and URL to buy it from. Then, it creates an Excel file
acording to the fields from the order format and the data from the URL
provided for each item.